feat(session): record csa_version + tool_version in session state (#714)
Sessions now persist the CSA binary version (`csa_version`, from
env!("CARGO_PKG_VERSION")) on creation and a best-effort tool version
(`tool_version` under each `[tools.<name>]`) captured at tool init
via a `--version` probe. Both fields are Option<String> and default
to None for backward compatibility — pre-feature session state files
deserialize silently.

`csa session list --csa-version <X>` filters sessions by recorded
binary version. More granular queries (range filters, tool-version
filter, sort-by) and the companion bisect-regression skill are
deferred to followups.

Closes #714.
