+++ crates/cli-sub-agent/src/pipeline_tests_post_exec_audit.rs
@@ -150,6 +150,7 @@ fn pre_execution_audit_baseline_returns_none_for_legacy_sessions_without_snapsho
         branch: None,
         created_at: chrono::Utc::now(),
         last_accessed: chrono::Utc::now(),
+        csa_version: None,
         genealogy: Default::default(),
         tools: Default::default(),
         context_status: Default::default(),
@@ -211,6 +212,7 @@ fn pre_execution_audit_baseline_prefers_per_execution_snapshot() {
         branch: None,
         created_at: chrono::Utc::now(),
         last_accessed: chrono::Utc::now(),
+        csa_version: None,
         genealogy: Default::default(),
         tools: Default::default(),
         context_status: Default::default(),
@@ -359,6 +361,7 @@ fn audit_failure_does_not_fail_execution() {
         branch: None,
         created_at: chrono::Utc::now(),
         last_accessed: chrono::Utc::now(),
+        csa_version: None,
         genealogy: Default::default(),
         tools: Default::default(),
         context_status: Default::default(),
@@ -474,6 +477,7 @@ fn reused_session_audit_uses_per_execution_baseline_not_session_creation() {
         branch: None,
         created_at: chrono::Utc::now(),
         last_accessed: chrono::Utc::now(),
+        csa_version: None,
         genealogy: Default::default(),
         tools: Default::default(),
         context_status: Default::default(),
@@ -568,6 +572,7 @@ fn first_execution_falls_back_to_session_creation_baseline_when_per_exec_capture
         branch: None,
         created_at: chrono::Utc::now(),
         last_accessed: chrono::Utc::now(),
+        csa_version: None,
         genealogy: Default::default(),
         tools: Default::default(),
         context_status: Default::default(),
