+++ crates/csa-executor/src/transport_meta.rs
@@ -630,6 +630,7 @@ mod tests {
             branch: None,
             created_at: now,
             last_accessed: now,
+            csa_version: None,
             genealogy: csa_session::state::Genealogy {
                 parent_session_id: None,
                 depth: 0,
