+++ crates/csa-executor/src/transport_tests_tail.rs
@@ -177,6 +177,7 @@ fn test_acp_build_env_propagates_extra_env() {
         branch: None,
         created_at: now,
         last_accessed: now,
+        csa_version: None,
         genealogy: csa_session::state::Genealogy {
             parent_session_id: None,
             depth: 0,
@@ -221,7 +222,6 @@ fn test_acp_build_env_propagates_extra_env() {
     );
 }
 
-
 #[test]
 fn test_acp_build_env_includes_csa_session_dir() {
     let transport = AcpTransport::new("claude-code", None);
@@ -233,6 +233,7 @@ fn test_acp_build_env_includes_csa_session_dir() {
         branch: None,
         created_at: now,
         last_accessed: now,
+        csa_version: None,
         genealogy: csa_session::state::Genealogy {
             parent_session_id: None,
             depth: 0,
@@ -295,6 +296,7 @@ fn test_acp_build_env_reserved_session_paths_override_extra_env() {
         branch: None,
         created_at: now,
         last_accessed: now,
+        csa_version: None,
         genealogy: csa_session::state::Genealogy {
             parent_session_id: None,
             depth: 0,
@@ -321,7 +323,10 @@ fn test_acp_build_env_reserved_session_paths_override_extra_env() {
     };
 
     let mut extra = HashMap::new();
-    extra.insert("CSA_SESSION_DIR".to_string(), "/tmp/fake-session".to_string());
+    extra.insert(
+        "CSA_SESSION_DIR".to_string(),
+        "/tmp/fake-session".to_string(),
+    );
     extra.insert(
         csa_session::RESULT_TOML_PATH_CONTRACT_ENV.to_string(),
         "/tmp/fake-session/result.toml".to_string(),
@@ -361,6 +366,7 @@ fn test_resume_session_id_extraction() {
         last_action_summary: String::new(),
         last_exit_code: 0,
         updated_at: now,
+        tool_version: None,
         token_usage: None,
     };
     let resume_id = tool_state.provider_session_id.as_deref();
@@ -375,6 +381,7 @@ fn test_resume_session_id_none_when_absent() {
         last_action_summary: String::new(),
         last_exit_code: 0,
         updated_at: now,
+        tool_version: None,
         token_usage: None,
     };
     let resume_id = tool_state.provider_session_id.as_deref();
@@ -443,6 +450,7 @@ fn build_test_meta_session(project_path: &str) -> MetaSessionState {
         branch: None,
         created_at: now,
         last_accessed: now,
+        csa_version: None,
         genealogy: csa_session::state::Genealogy {
             parent_session_id: None,
             depth: 0,
@@ -670,131 +678,4 @@ async fn test_execute_stops_after_max_attempts_and_returns_last_failure() {
     );
 }
 
-#[test]
-fn test_should_retry_gemini_rate_limited_until_final_attempt() {
-    let transport = LegacyTransport::new(Executor::GeminiCli {
-        model_override: None,
-        thinking_budget: None,
-    });
-    let execution = ExecutionResult {
-        summary: "failed".to_string(),
-        output: String::new(),
-        stderr_output: "HTTP 429 Too Many Requests".to_string(),
-        exit_code: 1,
-            peak_memory_mb: None,
-    };
-
-    assert!(
-        transport
-            .should_retry_gemini_rate_limited(&execution, 1, None)
-            .is_some()
-    );
-    assert!(
-        transport
-            .should_retry_gemini_rate_limited(&execution, 2, None)
-            .is_some()
-    );
-    assert!(
-        transport
-            .should_retry_gemini_rate_limited(&execution, 3, None)
-            .is_none()
-    );
-}
-
-#[test]
-fn test_should_not_retry_on_success_exit_code() {
-    let transport = LegacyTransport::new(Executor::GeminiCli {
-        model_override: None,
-        thinking_budget: None,
-    });
-    let execution = ExecutionResult {
-        summary: "ok".to_string(),
-        output: "429".to_string(),
-        stderr_output: String::new(),
-        exit_code: 0,
-            peak_memory_mb: None,
-    };
-    assert!(
-        transport
-            .should_retry_gemini_rate_limited(&execution, 1, None)
-            .is_none()
-    );
-}
-
-#[test]
-fn test_should_retry_on_quota_exhausted_marker() {
-    let transport = LegacyTransport::new(Executor::GeminiCli {
-        model_override: None,
-        thinking_budget: None,
-    });
-    let execution = ExecutionResult {
-        summary: "failed".to_string(),
-        output: String::new(),
-        stderr_output: "reason: 'QUOTA_EXHAUSTED'".to_string(),
-        exit_code: 1,
-            peak_memory_mb: None,
-    };
-    assert!(
-        transport
-            .should_retry_gemini_rate_limited(&execution, 1, None)
-            .is_some()
-    );
-}
-
-#[test]
-fn test_no_flash_fallback_stops_retry_after_attempt_2() {
-    let transport = LegacyTransport::new(Executor::GeminiCli {
-        model_override: None,
-        thinking_budget: None,
-    });
-    let execution = ExecutionResult {
-        summary: "failed".to_string(),
-        output: String::new(),
-        stderr_output: "HTTP 429 Too Many Requests".to_string(),
-        exit_code: 1,
-            peak_memory_mb: None,
-    };
-    let mut env = HashMap::new();
-    env.insert("_CSA_NO_FLASH_FALLBACK".to_string(), "1".to_string());
-    // Attempt 1 retries (advances to phase 2: same model + API key)
-    assert!(
-        transport
-            .should_retry_gemini_rate_limited(&execution, 1, Some(&env))
-            .is_some()
-    );
-    // Attempt 2 does NOT retry (phase 3 = flash, which is forbidden by no_flash)
-    assert!(
-        transport
-            .should_retry_gemini_rate_limited(&execution, 2, Some(&env))
-            .is_none()
-    );
-    // Without the flag, attempt 2 would still retry (advances to phase 3: flash)
-    assert!(
-        transport
-            .should_retry_gemini_rate_limited(&execution, 2, None)
-            .is_some()
-    );
-}
-
-#[test]
-fn test_no_failover_stops_retry_after_attempt_1() {
-    let transport = LegacyTransport::new(Executor::GeminiCli {
-        model_override: None,
-        thinking_budget: None,
-    });
-    let execution = ExecutionResult {
-        summary: "failed".to_string(),
-        output: String::new(),
-        stderr_output: "HTTP 429 Too Many Requests".to_string(),
-        exit_code: 1,
-        peak_memory_mb: None,
-    };
-    let mut env = HashMap::new();
-    env.insert("_CSA_NO_FAILOVER".to_string(), "1".to_string());
-
-    assert!(
-        transport
-            .should_retry_gemini_rate_limited(&execution, 1, Some(&env))
-            .is_none()
-    );
-}
+include!("transport_tests_tail_retry.rs");
