+++ crates/csa-scheduler/src/session_reuse.rs
@@ -136,6 +136,7 @@ mod tests {
             branch: None,
             created_at: chrono::Utc::now(),
             last_accessed: chrono::Utc::now(),
+            csa_version: None,
             genealogy: Default::default(),
             tools: Default::default(),
             context_status: Default::default(),
@@ -172,6 +173,7 @@ mod tests {
             branch: None,
             created_at: chrono::Utc::now(),
             last_accessed: chrono::Utc::now(),
+            csa_version: None,
             genealogy: Default::default(),
             tools: Default::default(),
             context_status: Default::default(),
@@ -211,6 +213,7 @@ mod tests {
             branch: None,
             created_at: chrono::Utc::now(),
             last_accessed: chrono::Utc::now(),
+            csa_version: None,
             genealogy: Default::default(),
             tools: Default::default(),
             context_status: Default::default(),
@@ -250,6 +253,7 @@ mod tests {
             branch: None,
             created_at: chrono::Utc::now(),
             last_accessed: chrono::Utc::now(),
+            csa_version: None,
             genealogy: Default::default(),
             tools: Default::default(),
             context_status: Default::default(),
@@ -290,6 +294,7 @@ mod tests {
             branch: None,
             created_at: old_time,
             last_accessed: old_time,
+            csa_version: None,
             genealogy: Default::default(),
             tools: Default::default(),
             context_status: Default::default(),
